Consolidate custom command scripts
The modtidy and buildvet scripts are now project agnostic.
Consolidated copies of semaphore.js, reverting it to its CommonJS
format from commit history.
